shared/gatt: Add bt_send_att_with_id functions
In case of char or descriptors search it might happen we have to send
couple of att commands and it is done under the hood.
In the same time used doing char discovery or descriptor search has
request id from the first ATT command, which means client gets outdated
if there is more command send. That means that client is not able to
e.g. cancle ongoing search if he would like to do it.